requirements.txt
loguru
redis
asyncpg
sqlalchemy
aiogram
Babel
python-dotenv

